FreeBSD portage.
git-svn-id: https://svn.janelia.org/eddylab/eddys/src/trunk/hmmer@695 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ testsuite/masks_test.c
@@ -3,7 +3,7 @@
  * 
  * Test driver for sequence masking routines in masks.c
  * 
- * RCS $Id$
+ * CVS $Id$
  */
 
 #include <stdio.h>
@@ -90,7 +90,7 @@ main(int argc, char **argv)
   /* XNU test
    */
   seq = test1;
-  len = strlen(seq);
+  len = (int) strlen(seq);
   dsq = DigitizeSequence(seq, len);
   XNU(dsq, len);
   result = MallocOrDie(sizeof(char) * (len+1));
